Change Craft Ordering Quantity to Allow Longs (#672)

+++ src/main/java/appeng/client/gui/implementations/GuiPatternMulti.java
@@ -128,7 +128,7 @@ public class GuiPatternMulti extends GuiAmount {
     }
 
     @Override
-    protected int addOrderAmount(final int i) {
+    protected long addOrderAmount(final int i) {
         return i + getAmount();
     }
